Rename `cp` to `cpoptimizer` module
