+++ pyjobshop/cpoptimizer/result2solution.py
