+++ pyjobshop/cpoptimizer/variables.py
